task_planner_app/requirements/requirements.txt
pytz~=2022.1
argon2-cffi~=21.3.0 #argon password hasher
django-widget-tweaks~=1.4.12
django-filter~=21.1
django-settings-export~=1.2.1
python-json-logger~=2.0.2
django-bootstrap-breadcrumbs==0.9.2
pandas==1.1.5
pillow==9.1.1
# Django
# -----------------------------------------------------
django==3.2.13
django-model-utils==4.2.0
django-crispy-forms==1.14.0
django-select2==7.10.0
django-braces==1.15.0
